fix(clusters): Fix diffing clusters
Because clusters have some properties that are JSON array and the order is arbitrary, it does not make sense to hash the JSON of the model. Instead I've brought in the Javers library which looks like it has some pretty exciting potential to use to persist historical diffs, provide more context to users as to why Keel is stomping their resources, and for Keel to make smarter decisions about annealing based on what exactly has changed.