feat(cluster): cluster converge with unhealthy instances
Changes the cluster task to use the rollback on failure flag, and times out the stage faster (30mins instead of 2hrs).

Rollback on failure only applies to instances failing to come up healthy, this change attempts to leave the cluster in
a previous-known-good but also needs-to-reconverge state so that we will get retries on deployment via keel attempting
to converge the cluster repeatedly.