fix(security groups): require vpc on locations but default to vpc0
This fixes the situation where we can't resolve current state of a security group because its `vpc` property is `null`. We need to map `vpc` to a VPC id so we can send that to CloudDriver to fetch the current state. We don't currently apply resolvers to the spec passed to `ResourceHandler.desired` so the property can be `null` if the spec relies on the default applied by `NetworkResolver`. This is an interim fix until we figure out the proper approach.