refactor(bakery): made BaseImageCache an interface
The implementation (which is reliant on Mahe) will move into `keel-nflx` as it will be no use to open-source users anyway.
I discussed with @asher the possibility of getting base AMI versions from the Rosco bake options endpoint, which I think will become the default implementation. Ideally we could also add the same endpoint to the Netflix bakery and obviate the need for Mahe.