fix(clouddriver): healthCheckPort as string
When we query clouddriver for an ApplicationLoadBalancer (ALB) model object,
one of the model sub-objects, TargetGroup, has a field named
`healthCheckPort`.

This field is encoded as a string, but is almost always a string-quoted
integer (e.g., `"8080"`). However, a valid value is the string
`"traffic-port"`.

Keel's TargetGroup model object represents this value as an `Int`, which
leads to a deserialization error when deserializing an ALB with this
value.

To fix this, we need to modify the `TargetGroup` model so that it can deserialize
ALB models that have health check port set to either a number (e.g.,
"8080") or the literal string "traffic-port".

I changed the field on the ApplicationLoadBalancer model object from `Int` to `String`.

However, I did **not** modify the corresponding field on the
ApplicationLoadBalancerSpec.TargetGroup object.

Instead I modified the `CloudDriverService.getApplicationLoadBalancer` method in
ApplicationLoadBalancerHandler.kt to convert
"traffic-port" to the numerical value of the traffic port.

Rationale: I thought that preserving the current data model of the spec
object would be less risky: Since the spec objects get
persisted to the database, making changes to them carries a risk of the
code and data going out of sync.