fix(actuator): check for resource by id
We are seeing an issue in deployment where DeliveryConfig.environmentFor seems to
be returning null, even though the resource exists in an environment in
the delivery config.

One hypothesis is that the `contains` check is not properly matching a
resource object.

This code changes the check so it checks by resource id string, rather
than by using the equality check.