fix(ec2): gracefully handle dangling security group ref
Context

Every AWS security group specifies a set of *inbound rules*. There
are different types of inbound rules. One particular type,
"a security group rule reference", refers to a second security group.
Essentially, it says "Allow instances in that security group to talk
to me."

Problem

Generally, AWS does not allow a user to delete a security group if other
security groups refer to it in an inbound rule. However, there are
edge cases where this can happen. As a result, a security group might
have an inbound rule that refers to another security group that no
longer exists.

If you query clouddriver for information about a security group that has
inbound rules that point to another security group that no longer
exists, the referenced security group will have a missing "name" field
in the payload, for example:

```json
{
 "class": "com.netflix.spinnaker.clouddriver.aws.model.AmazonSecurityGroup",
  "inboundRules": [
    {
      "class": "com.netflix.spinnaker.clouddriver.model.securitygroups.SecurityGroupRule",
      "securityGroup": {
        "class": "com.netflix.spinnaker.clouddriver.aws.model.AmazonSecurityGroup",
	"id": "sg-xxxxxxxxxxxxxxxxx" <--- sg that no longer exists
	// Missing "name" field
```

When keel tries to deserialize the corresponding
`com.netflix.spinnaker.keel.clouddriver.model.SecurityGroupModel`
object, it throws an exception because the `name` field is required:

```
Instantiation of [simple type, class com.netflix.spinnaker.keel.clouddriver.model.SecurityGroupModel$SecurityGroupRuleReference] value failed for JSON property name due to missing (therefore NULL) value for creator parameter name which is a non-nullable type
 at [Source: (okhttp3.ResponseBody$BomAwareReader); line: 1, column: 1317] (through reference chain: com.netflix.spinnaker.keel.clouddriver.model.SecurityGroupModel["inboundRules"]->java.util.HashSet[1]->com.netflix.spinnaker.keel.clouddriver.model.SecurityGroupModel$SecurityGroupRule["securityGroup"]->com.netflix.spinnaker.keel.clouddriver.model.SecurityGroupModel$SecurityGroupRuleReference["name"])
```

Proposed solution

Allow the `SecurityGroupRuleReference.name` field to be null to avoid
the deserialization error.

In the SecurityGroupHandler logic, when converting from
`com.netflix.spinnaker.keel.clouddriver.model.SecurityGroupModel` to `com.netflix.spinnaker.keel.api.ec2.SecurityGroup`,
if the name field is null, then omit that inbound rule.

This should not affect behavior since an inbound rule that points to a non-existent security group has no effect.

Other changes

We now log a warning when encountering an inbound rule to a non-existent
security group. This log includes the invalid security group id, to make it easier for an operator to find this rule
if they wanted to look for it.