feat(verify): only promote on verif success
This is work towards supporting verification.

Problem

We don't want to promote a version to the next environment if there are verifications that haven't succeeded yet.

Proposed solution

Extend the behavior of the DependsOnConstraintEvaluator to enforce that all verifications have passed in the required environment before allowing promotion.

The DependsOnConstraintEvaluator seemed like a natural place to put this logic.

Future work

Provide feedback to the end user about the status of verifications. This
code doesn't provide any user/operator feedback about the status of
verifications. That will likely live elsewhere in the codebase.