feat(api): start to add verification info
## Context

Deck makes the following API call to populate the environments view:

```
/application/{application}?entities=resources&entities=artifacts&entities=environments
```

## Problem

The above API call needs to expose information about the state of verifications in order for Deck to display information about the state of verifications in the environments view.

## Work in this PR

The above API call returns a list of environments when `entities=environments` is specified.  With this PR, each entry in the list contains information about which verifications are specified in the delivery config.

Here's example (partial) response:

```json
"environments": [
  {
   "name": "staging",
   ...
   "verifications": [
     {
      "type": "test-container",
      "id": "test-container:acme/end-to-end-tests:stable"
     }
   ]
  },
  ...
]
```

## Future work

In a future PR, the `artifacts` section of the response will be augmented to show information about the state of verifications (e.g., running, passed, failed).

I decided to break up the work in two pieces to make the PRs easier to review.