feat(verif): more robust handling of invalid verification id
+ Problem

Environments view was blowing up for a test app in our staging environment because the `verification_state` table had an invalid verification id in one of the entries.

The invalid data looked like this: `spkr/mytest/stable` where the valid id was `spkr/mytest:stable`.

We had changed the convention for the id a while back from `/` to `:` for delimiting the tag of the container, but there was some data that hadn't been updated.

While this could have been fixed by just repairing the data in the database, having the environments view break entirely because of one invalid verificiations record is an undesirable user experience.

+ Implemented solution

The `VerificationContext.verification(id: String)` function now returns a null if the id doesn't exist, which forces the caller to explicitly handle it.

++ Environment view

The ApplicationService, which computes the environments view, now
handles this by skipping the invalid record. It also logs an error
message and emits a metric: `keel.verification.invalid.id.seen`

++ Verification controller

The `VerificationController` returns a meaningful error message if a client calls it with an invalid id. Here's example output

```yaml
timestamp: "2021-02-11T17:51:50.148+0000"
status: 404
error: "Not Found"
message: "Unknown verification id: spkr/fakeetest:fail. Expecting one of: [spkr/faketest:fail]"
```