fix(api): don't blow up the api if can't query for verifications
# Problem

We hit an issue (documented as as #1821) where the batch verification query was throwing an exception, which broke the view.

# Implemented solution

Degrade gracefully by catching the exception, logging the error, and behaving as if there are no verifications.

This PR doesn't actually fix the bug (that's coming later), but it ensures that the UI doesn't break just because we can't fetch verification info.