fix(envts): initial work to fix verification/deployment races
## Summary

This is the first of a series of PRs to squish some race conditions.

This PR starts an a solution at the high-level, with details to be filled in in future PRs.

There are no (intended) behavior changes in this PR.

## Context

Here are two [safety properties][1] of the verification behavior of Managed Delivery.

**P1: Two verifications should never execute concurrently against the same environment.**

For example, if the `acme/tests:stable` test container is currently running in the `staging` environment, keel should
not launch any other verifications in the `staging` environment.

**P2: Verifications against an environment should never happen concurrently with a deployment in an environment.**

For example, if the `fnord-v123.deb` artifact is being deployed to an EC2 cluster in the `staging` environment, keel should not launch verifications while this is happening.

## Problem

Because of race conditions, we don't properly enforce these safety properties today. See #1830, #1831.

## Sketch of overall solution

Add guard conditions before launching a verification or an actuation.

While it may not be strictly necessary to guard against *all* actuations, it simplifies the implementation to guard actuations instead of figuring out which actuations correspond to deployments.

### Enforcer and leases

These guard conditions are enforced by a new class named `EnvironmentExclusionEnforcer`.

The enforcer works by granting the client a *lease* if the guard conditions are met.
A lease is a lock that has an expiration time.
Leases expire to protect against the situation where an instance takes a lease and then terminates unexpectedly before releasing it.

Lease enforcement (including expiration) will be implemented in a future PR.

### Verification guard conditions

Before starting a verification in an environment E, ensure that:
* No other verifications are currently running in E
* No deployments are currently happening in E

Implemented by: `EnvironmentExclusionEnforcer.withVerificationLease`

### Actuation guard conditions

Before starting an actuation in an environment E, ensure that:

* No other verifications are running in E

Implemented by: `EnvironmentExclusionEnforcer.withActuationLease`

## Work in this PR

* Create the `EnvironmentExclusionEnforcer` class
* Call the enforcer before launching a verification
* Call the enforcer before actuating a resource

All of the actual checks are no-ops and will be implemented in a future PR.

[1]: https://buttondown.email/hillelwayne/archive/safety-and-liveness-properties/