fix(resource): don't show error if verification prevents actuation
## Problem

If a resource was prevented from being actuation because there was a verification running in the environment, we show the user an error in the event history for the resource. We shouldn't do this, because this isn't really an error state, it's something that will resolve on its own once the verification completes.

There was a bug in the code where the related exception was getting handled further down in the stack than expected.

## Implemented solution

Move the exception-handling logic so that it catches the exception that gets thrown when actuation fails due to verification. Don't treat this as an error anymore.

We also now increment a metric when this happens: `keel.actuation.blocked`

This metric is tagged with resourceId, resourceKind, and resourceApplication.