fix(test): fix occasional failing sql test
## Problem

Occasionally, we see an SQL test failure that looks like this:

SqlCombinedRepositoryTests individual resources are persisted FAILED

  org.jooq.exception.DataAccessException: SQL [insert into `keel`.`event` (`uid`, `scope`, `json`) values (?, ?, ?)]; Incorrect datetime value: '2021-04-03T19:06:08Z' for function str_to_date

  Caused by: java.sql.SQLException: Incorrect datetime value: '2021-04-03T19:06:08Z' for function str_to_date

This happens when the fractional second component of the serialized Instant object is *exactly* zero. In that case, the string representation doens't have a fractional second component. This causes the mysql str_to_date function to fail, because it expects a fractional component to be there.

## Implemented solution

In the MutableClock class, which is used for generating Instant values in the tests, ensure that the returned Instant objects never has a zero nanosecond component.

This could potentially manifest in prod, although it's less likely, and we've never actually seen it.

Logner term, we could fix this by writing a custom serializer for Instant to ensure it writes out a fractional component of `.000000` instead of leaving it out.