fix(verif): set pending to override fail if verification removed
Problem

If a running verification V in environment E gets "stuck" (i.e., runs forever), and the user removes V from the delivery config, keel will still think that V is
running in E, and which will prevent new deployments.

Keel checks if a verification is running in the environment by querying the action_state table, and a stuck verification still has a PENDING record in that table.

Implemented solution

When the delivery config is updated, check if a verification has been removed. If so, set all of the relevant PENDING entries in action_state to OVERRIDE_FAIL.

Squashed commit of the following:

commit c0ed99deb774652e93cf6c8c8e635fe2d7254cd0
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Thu Jun 10 17:10:43 2021 -0700

    Update to OVERRIDE_FAIL instead of deleting

commit 526eee6fe864abd85de45b779c45b8f82a7027c6
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Wed Jun 9 14:34:09 2021 -0700

    fix(verif): clear state associated with removed verification

    Problem:

    If a running verification V in environment E gets "stuck" (i.e., runs forever), and the
    user removes V from the delivery config, keel will still think that V is
    running in E, and which will prevent new deployments.

    Keel checks if a verification is running in the environment by querying
    the action_state table, and a stuck verification still has a PENDING
    record in that table.

    Implemented solution:

    When the delivery config is updated, check if a verification has been
    removed. If so, remove the records in the action_state table associated
    with the removed verification.

    Questions:

    Is it reasonable to delete data from the database like this? Would it be better
    to update the record so it's not marked as PENDING anymore? (e.g., mark
    it as OVERRIDE_FAIL). I feel like it's cleaner to remove orphaned data
    from the database. But it means that if a user removes a verification
    and then adds it back, they'll lose their history.

(cherry picked from commit a1f06521c649692130f8aba346c39493aaa170cb)